Move arrow/compute/kernels/registry.cc to arrow/compute/initialize.cc and arrow/compute/kernels/api.h to arrow/compute/initialize.h. Add it to arrow/compute/api.h
