+++ cpp/examples/arrow/compute_and_write_csv_example.cc
@@ -17,7 +17,6 @@
 
 #include <arrow/api.h>
 #include <arrow/compute/api.h>
-#include <arrow/compute/kernels/api.h>
 #include <arrow/csv/api.h>
 #include <arrow/csv/writer.h>
 #include <arrow/io/api.h>
