+++ cpp/examples/arrow/join_example.cc
@@ -23,7 +23,6 @@
 #include <arrow/csv/api.h>
 #include "arrow/acero/exec_plan.h"
 #include "arrow/compute/expression.h"
-#include "arrow/compute/kernels/api.h"
 
 #include <arrow/dataset/dataset.h>
 #include <arrow/dataset/plan.h>
