+++ cpp/examples/tutorial_examples/dataset_example.cc
@@ -19,7 +19,7 @@
 
 // (Doc section: Includes)
 #include <arrow/api.h>
-#include <arrow/compute/kernels/api.h>
+#include <arrow/compute/api.h>
 #include <arrow/dataset/api.h>
 // We use Parquet headers for setting up examples; they are not required for using
 // datasets.
