+++ cpp/src/arrow/compute/initialize.h
