+++ cpp/src/arrow/compute/kernels/test_util_internal.h
@@ -28,8 +28,8 @@
 #include "arrow/array.h"
 #include "arrow/compute/api_scalar.h"
 #include "arrow/compute/cast.h"
+#include "arrow/compute/initialize.h"
 #include "arrow/compute/kernel.h"
-#include "arrow/compute/kernels/api.h"
 #include "arrow/compute/test_util_internal.h"
 #include "arrow/datum.h"
 #include "arrow/memory_pool.h"
