+++ python/pyarrow/includes/libarrow.pxd
@@ -2228,6 +2228,8 @@ cdef extern from "arrow/util/thread_pool.h" namespace "arrow::internal" nogil:
 
 cdef extern from "arrow/compute/api.h" namespace "arrow::compute" nogil:
 
+    CStatus InitializeCompute " arrow::compute::Initialize"()
+
     cdef cppclass CExecBatch "arrow::compute::ExecBatch":
         vector[CDatum] values
         int64_t length
@@ -2884,9 +2886,6 @@ cdef extern from "arrow/compute/api.h" namespace "arrow::compute" nogil:
         const shared_ptr[CScalar]& scalar() const
 
     cdef c_string ToString(DatumType kind)
-
-cdef extern from "arrow/compute/kernels/api.h" namespace "arrow::compute" nogil:
-    CStatus InitializeCompute " arrow::compute::Initialize"()
 
 
 cdef extern from * namespace "arrow::compute":
